Conflicts:
	imageplugins/enhance/digikamimageplugin_enhance_ui.rc
	tests/CMakeLists.txt

+++ imageplugins/enhance/sharpentool.cpp
@@ -6,7 +6,7 @@
  * Date        : 2004-07-09
  * Description : a tool to sharp an image
  *
- * Copyright (C) 2004-2012 by Gilles Caulier <caulier dot gilles at gmail dot com>
+ * Copyright (C) 2004-2014 by Gilles Caulier <caulier dot gilles at gmail dot com>
  *
  * This program is free software; you can redistribute it
  * and/or modify it under the terms of the GNU General
@@ -185,7 +185,7 @@ void SharpenTool::preparePreview()
         case SharpContainer::UnsharpMask:
         {
             DImg img  = d->previewWidget->getOriginalRegionImage();
-            int    r  = (int)settings.umRadius;
+            double r  = settings.umRadius;
             double a  = settings.umAmount;
             double th = settings.umThreshold;
 
@@ -243,7 +243,7 @@ void SharpenTool::prepareFinal()
 
         case SharpContainer::UnsharpMask:
         {
-            int    r  = (int)settings.umRadius;
+            double r  = settings.umRadius;
             double a  = settings.umAmount;
             double th = settings.umThreshold;
